Refactor to use ES6 Modules
