+++ index.html
@@ -18,6 +18,6 @@
       <input type="text" id="author" placeholder="Author" /><br />
       <button id="add-book" type="submit">Add</button>
     </form>
-    <script src="index.js"></script>
+    <script type="module" src="js/index.js"></script>
   </body>
 </html>
